+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.5.0] - 2024-04-18
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Contact us form page.
-- [Marek Kurańda](https://github.com/mjkuranda): `InputString` component.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputString` and `InputTextarea` components.
 
 ## [0.4.0] - 2024-04-17
 ### Added